Update Guidelines Incipit and Explicit
I updated the guidelines concerning the incipits and explicits as discussed recently in our latest webmeeting. I am not sure, whether the phrase on colophons is useful or should deleted or extended. It might be appropriate to refer to the subpage "colophon", but I did not know, how I can indicate it in a ref.