Update Guidelines Incipit and Explicit (#142)
* Update Guidelines Incipit and Explicit

I updated the guidelines concerning the incipits and explicits as discussed recently in our latest webmeeting. I am not sure, whether the phrase on colophons is useful or should deleted or extended. It might be appropriate to refer to the subpage "colophon", but I did not know, how I can indicate it in a ref.

* Updated definition of Implicit and explicit

I tried to include Denis' and Dorotheas quests.

* New suggestion for explicit and incipit

One more suggestion to turn the perspective to make it sound less strict (both in incipit and explicit).

* Update explicit.xml

* Update explicit.xml according to Eugenias request

* Update incipit/explicit

I changed the records according to Denis request and rewrote some other phrases to make it a bit smoother.