Update Guideline (replace 'texual unit')
I changed the term "textual unit" to "Work" or to "Work or Narrative Unit" to make it more clear according to our discussion in https://github.com/BetaMasaheft/Documentation/issues/1918#issuecomment-3009520168. I kept the "textual units" only in a few cases as for instance in create.xml in the paragraph on text identification (Example workflow) because I felt in this case, that the term is justified since it refers to a case, when it is not certainly clear whether a text can be identified as Work, Narrative unit or none of that.

I updated some terms to capitalised terms, because I felt, that this is the dominant method. However, I did not do this systematically in all records in this pull request.

I have commented out some details, that possibly need some modification in other pull requests.